backend/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-JWT-Extended==4.7.*
Flask-Cors==5.0.*
Flask-Limiter==3.8.*
marshmallow==3.23.*
python-dotenv==1.1.*
requests==2.32.*
openai==1.109.*
gunicorn==23.*
